+++ webpack.config.js
@@ -16,7 +16,7 @@ const coreConfig = (env = {}) => ({
         contentExtension: 'home',
         additionalContent: [
           {
-            title: 'Home',
+            label: 'Home',
             filePath: path.resolve(process.cwd(), 'README.md'),
           },
         ],