+++ packages/terra-application-navigation/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
+## 1.55.0 - (September 20, 2021)
+
 * Changed
+  * Updated behavior for the menu close custom event to trigger onDrawerMenuStateChange callback.
+  * Updated WDIO screenshots due to webpack-config-terra sass upgrade.
   * Updated WDIO snapshots due to terra-functional-testing v2 upgrade.
 
 ## 1.54.0 - (July 21, 2021)