+++ packages/terra-theme-provider/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+## 4.10.0 - (September 20, 2021)
+
 * Changed
+  * Updated WDIO screenshots due to webpack-config-terra sass upgrade.
   * Updated fusion wdio tests to align with other tests
 
 ## 4.9.0 - (June 8, 2021)