+++ packages/terra-time-input/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+## 4.43.0 - (September 20, 2021)
+
 * Changed
+  * Updated WDIO screenshots due to webpack-config-terra sass upgrade.
   * Improved specificity of meridiem button group.
   * Check NODE_ENV for the current development environment.